Added text bar, todo list and solved offset problem
Added the text bar in the bottom and parameterized some messages when
adding objects to the canvas.
Solved the offset problem on the canvas, putting a selection of the
offset to make according to the OS.
Also added a todo list, so we can keep track of tasks to do.