Implementation of the changes suggested in the assignment discussion
Implementation of the changes suggested by the professor during the
assignment discussion, namely the creation of an interface for code
generation, in the context menu item plug-in. Now there are two
different interfaces: one that is demanding to be implemented and an
optional one that is implemented only in case the plug-in generated any
source code.
All the extension point interfaces were documented, as well as the
plug-in schemas